refactor(browser): move persistence transforms to shared module
